Split JS interop libraries into separate files
